refactor(files): seperate files
